Add initial sandboxing tools
sandbox.py lets you create a .aci file (App Container Image) from a tar
file, that should run a given command when executed.

exec/chroot.py executes a .aci file (App Container Image).

I doubt that the produced .aci files are actually compatible with other
App Container ('appc') execution tools yet.

The executor will also be unable to run all but the simplest
applications that are packaged as .aci files.

Also, the process of untarring a .tar file to create a .aci, then
unpacking the .aci again to run it is very ineffecient.

Plus it is missing lots of important features.

But, it's a start.