Change how we search for 'linux-user-chroot'
First, the user should get better errors now if it is not found:

    sandboxlib.ProgramNotFound: Did not find 'linux-user-chroot' in
    PATH. Searched '/sbin:/bin:/usr/sbin:/usr/bin'

Second, we explicitly search for the program on each call to
run_sandbox() using the same search code used in
sandbox_module_for_platform(). This shouldn't change anything, but I
think it's better practice to search for the program ourselves than to
assume exec('linux-user-chroot') will do the right thing.