Catch OSError when we attempt to run which
Some distributions define which as an alias,
for example in NixOS which is defined as `type -P',
since subprocess isn't running a shell the which cmd is not found
and an OSError is raised, it's probably reasonable to catch this,
since if we can't find which we can't find linux-user-chroot and
should fallback to chroot. We could run a shell, but it's probably
not worth the security hazard.