sandboxlib.linux_user_chroot: Allow manually specified read-only mounts
You can't create a bind-mount as read-only,
you can only bind-mount then remount it as read-only.

So a sandboxlib user might opt to say it wants to bind something in,
then make it read-only, as two separate extra mounts.

We can't do this directly with linux-user-chroot,
as we are restricted to bind-mounts and making a subtree read-only,
but making a subtree read-only is close enough.