sandboxlib.chroot: Don't attempt to unmount remount mounts
The mount operation is overloaded to also remount to change flags.

This does not result in a new mount,
so unmounting it is the wrong thing to do in this case.

For now, we assume that we're modifying a mount we created earlier,
so we can just avoid unmounting when we remount,
rather than having to determine how to reverse the changing of flags.